+++ plugin-packs/postcss-preset-env/.tape.mjs
@@ -171,7 +171,7 @@ postcssTape(plugin)({
 		options: {
 			stage: 0,
 			browsers: '> 0%'
-		},
+		}
 	},
 	'layers-basic:preserve:true': {
 		message: 'supports layers usage with { preserve: true }',
@@ -179,7 +179,7 @@ postcssTape(plugin)({
 			preserve: true,
 			stage: 0,
 			browsers: '> 0%'
-		},
+		}
 	},
 	'client-side-polyfills:stage-1': {
 		message: 'stable client side polyfill behavior',