+++ plugins/postcss-custom-selectors/CHANGELOG.md
@@ -3,10 +3,29 @@
 ### Unreleased (major)
 
 - Updated: Support for Node v14+ (major).
+- Removed: `importFrom` feature (breaking).
+- Removed: `exportTo` feature (breaking).
+- Fixed: follow the specification and use `:is()` in transformed selectors (breaking).
+- Added: Support for `@scope` and `@container` as parent rules of `@custom-selector`.
+
+```diff
+@custom-selector :--heading h1, h2, h3;
+
+article :--heading + p {
+	margin-top: 0;
+}
+
+/* becomes */
+
+- article h1 + p,article h2 + p,article h3 + p {
++ article :is(h1, h2, h3) + p {
+	margin-top: 0;
+}
+```
 
 ### 6.0.3 (June 4, 2022)
 
-- Fixed: allow any valid ident in custom media (`@custom-selector :--🧑🏾‍🎤 .singer`)
+- Fixed: allow any valid ident in custom selectors (`@custom-selector :--🧑🏾‍🎤 .singer`)
 
 ### 6.0.2 (June 3, 2022)